HBX-1650: Move class 'org.hibernate.tool.hbm2x.Cfg2HbmTool' to 'org.hibernate.tool.internal.export.hbm.Cfg2HbmTool'
