+++ orm/src/main/java/org/hibernate/tool/internal/export/hbm/HibernateMappingExporter.java
@@ -10,7 +10,6 @@ import java.util.Map;
 import org.hibernate.boot.Metadata;
 import org.hibernate.internal.util.StringHelper;
 import org.hibernate.mapping.PersistentClass;
-import org.hibernate.tool.hbm2x.Cfg2HbmTool;
 import org.hibernate.tool.hbm2x.GenericExporter;
 import org.hibernate.tool.hbm2x.HibernateMappingGlobalSettings;
 import org.hibernate.tool.hbm2x.TemplateProducer;
